% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_fonts.R
\name{google_fonts}
\alias{google_fonts}
\title{Meta-information for the Google Fonts}
\usage{
google_fonts(name, ...)
}
\arguments{
\item{name}{Name of the font that will be searched in Google Fonts.}

\item{\dots}{Other parameters passed to \code{curl::curl_fetch_memory()}.}
}
\description{
This function provides information of Google Fonts
that can be used in the \code{\link{font_install}()} function.
It will try to include all font faces available for a specified family name.
}
\examples{
opensans = google_fonts("Open Sans")
print(opensans)

\dontrun{
## Install the Open Sans font to the showtexdb package
font_install(opensans)
}
}
\author{
Yixuan Qiu <\url{https://statr.me/}>
}
