% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_font.R
\name{install_font}
\alias{install_font}
\title{Install Fonts to the 'showtextdb' Package}
\usage{
install_font(font_desc, quiet = FALSE)
}
\arguments{
\item{font_desc}{A list that provides necessary information of the font
for installation. See the \strong{Details} section.}

\item{quiet}{Whether to show the progress of downloading and installation.}
}
\description{
This function saves the specified font to the \file{fonts} directory of the
\pkg{showtextdb} package, so that it can be used by the \pkg{showtext}
package.
}
\details{
\code{font_desc} should contain at least the following components:

\describe{
  \item{\code{showtext_name}}{The family name of the font that will be used in
                              \pkg{showtext}.}
  \item{\code{font_ext}}{Extension name of the font files, e.g., \code{ttf} for
                         TrueType, and \code{otf} for OpenType.}
  \item{\code{regular_url}}{URL of the font file for "regular" font face.}
}

Optionally, \code{font_desc} can also contain \code{bold_url}, \code{italic_url},
\code{bolditalic_url}, and \code{symbol_url} that provide the URLs of the other
font faces.

See \code{\link{source_han_sans}()} and \code{\link{source_han_serif}()}
for an example of the \code{font_desc} parameter.
}
\examples{
\dontrun{
install_font(source_han_serif())
}
}
\author{
Yixuan Qiu <\url{http://statr.me/}>
}
