% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fonts.R
\name{load_fonts}
\alias{load_fonts}
\title{Load Fonts for the 'showtext' Package}
\usage{
load_fonts()
}
\description{
This function loads fonts that will be used by the \pkg{showtext} package.
}
\examples{
\dontrun{
sysfonts::font.families()

load_fonts()
sysfonts::font.families()
}
}
\author{
Yixuan Qiu <\url{http://statr.me/}>
}
