% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_showtext_fonts.R
\name{load_showtext_fonts}
\alias{load_showtext_fonts}
\title{Load Fonts for the 'showtext' Package}
\usage{
load_showtext_fonts()
}
\description{
This function loads fonts that will be used by the \pkg{showtext} package.
}
\examples{
\dontrun{
sysfonts::font_families()

load_showtext_fonts()
sysfonts::font_families()
}
}
\author{
Yixuan Qiu <\url{https://statr.me/}>
}
