% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_han.R
\name{source_han}
\alias{source_han}
\alias{source_han_sans}
\alias{source_han_serif}
\title{Meta-information for the Source Han Sans/Serif Fonts}
\usage{
source_han_sans(lang = c("CN", "TW", "JP", "KR"))

source_han_serif(lang = c("CN", "TW", "JP", "KR"))
}
\description{
These functions provide information of the Source Han Sans/Serif fonts
that can be used in the \code{\link{install_font}()} function.
Source Han Sans/Serif fonts provide complete support for the
CJK (\strong{C}hinese, \strong{J}apanese, and \strong{K}orean) characters.
}
\examples{
\dontrun{
## Install Source Han Sans font (by default Simplified Chinese)
## to the showtexdb package
install_font(source_han_sans())

## Source Han Serif Japanese
install_font(source_han_serif("JP"))
}
}
\author{
Yixuan Qiu <\url{http://statr.me/}>
}
