% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_fonts.R
\name{load_fonts}
\alias{load_fonts}
\title{Loading Fonts For The 'showtext' Package}
\usage{
load_fonts()
}
\description{
This function loads fonts that will be used by the \pkg{showtext} package.
}
\examples{
load_fonts()
}
\author{
Yixuan Qiu <\url{http://statr.me/}>
}

